add & remove collection

+++ src/redux/features/collectionSlice.js
@@ -1,5 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
-
+import  { toast, Zoom } from "react-toastify";
+ 
 
 const initialState={
             items:JSON.parse(localStorage.getItem('collection'))||[]
@@ -9,16 +10,65 @@ const collectionSlice = createSlice({
             initialState,
             reducers:{
                         addCollection:(state,action)=>{
-                                    state.items.push(action.payloas);
-                                    localStorage.setItem('collection',JSON.stringify(state.items))
+                                    const alreadyExists = state.items.find(
+                                                item=>item.id == action.payload.id
 
+                                    )
+                               if(!alreadyExists){
+                                    state.items.push(action.payload);
+                                    localStorage.setItem('collection',JSON.stringify(state.items))
+                               }
                         },
                         removeCollection:(state,action)=>{
-                                    
+                                    state.items=state.items.filter(
+                                    item => item.id !== action.payload
+                                    )
+                                     localStorage.setItem('collection',JSON.stringify(state.items))
                         },
                         clearCollection:(state,action)=>{
+                                    state.items=[]
+                                    localStorage.removeItem('collection')
+                        },
+                        addedToast:()=>{
+                                    toast.success('Added to collection', {
+                                        position:"top-center",
+                                   autoClose:2000,
+                              hideProgressBar:false,
+                              closeOnClick:true,
+                              pauseOnHover:false,
+                              draggable:true,
+                              progress:undefined,
+                              theme:'dark',
+                              transition:Zoom,
+                         
+                         });
 
-                        }
-            }
+                        },
+                         removeToast:()=>{
+                                    toast.error('Remove from collection', {
+                                        position:"top-center",
+                                   autoClose:2000,
+                              hideProgressBar:false,
+                              closeOnClick:false,
+                              pauseOnHover:true,
+                              draggable:true,
+                              progress:undefined,
+                              theme:'dark',
+                              transition:Zoom,
+
+            });
+          }}
 
 })
+
+export const{
+            addCollection,
+            removeCollection,
+            clearCollection,
+            addedToast,
+            removeToast,
+            
+
+}=collectionSlice.actions;
+
+export default collectionSlice.reducer;
